requirements.txt
pip>=22.2.2
nibabel==4.0.2
numpy==1.23.3
PyYAML==6.0